+++ R/find.R
@@ -36,7 +36,7 @@
 #' (input <- tabs2shift$clusters_messy)
 #'
 #' schema <- setCluster(id = "territories",
-#'                      left = .find(pattern = "comm*"), top = .find(pattern = "comm*")) %>%
+#'                      left = .find(pattern = "comm*"), top = .find(pattern = "unit")) %>%
 #'   setIDVar(name = "territories", columns = c(1, 1, 4), rows = c(2, 9, 9)) %>%
 #'   setIDVar(name = "year", columns = 4, rows = c(3:6), distinct = TRUE) %>%
 #'   setIDVar(name = "commodities", columns = c(1, 1, 4)) %>%