+++ tests/testthat/test-08_groups.R
@@ -33,16 +33,16 @@ test_that("apply function to summarise merged rows", {
   input <- tabs2shift$group_sum
 
   schema <-
-    setGroups(rows = .sum(c(3, 4))) %>%
-    setGroups(rows = .sum(c(6, 7))) %>%
-    setIDVar(name = "territories", columns = 1) %>%
-    setIDVar(name = "year", columns = 2) %>%
-    setIDVar(name = "commodities", columns = c(3:6), rows = 2) %>%
-    setObsVar(name = "harvested", columns = c(3, 4)) %>%
+    setGroups(rows = .sum(c(3, 4))) |>
+    setGroups(rows = .sum(c(6, 7))) |>
+    setIDVar(name = "territories", columns = 1) |>
+    setIDVar(name = "year", columns = 2) |>
+    setIDVar(name = "commodities", columns = c(3:6), rows = 2) |>
+    setObsVar(name = "harvested", columns = c(3, 4)) |>
     setObsVar(name = "production", columns = c(5, 6))
 
-  reorganise(input = input, schema = schema) %>%
-    arrange(territories, year, commodities) %>%
+  reorganise(input = input, schema = schema) |>
+    arrange(territories, year, commodities) |>
     .expect_valid_table(units = 2)
 
 })